refactor: use analysis > RiskLevel
