--- src/main/java/com/blockguard/server/domain/fraud/domain/enums/RiskLevel.java
@@ -1,17 +1,0 @@
-package com.blockguard.server.domain.fraud.domain.enums;
-
-import com.fasterxml.jackson.annotation.JsonValue;
-import lombok.AllArgsConstructor;
-
-@AllArgsConstructor
-public enum RiskLevel {
-    Danger("위험"),
-    Safety("안전");
-
-    private final String value;
-
-    @JsonValue
-    public String getValue(){
-        return value;
-    }
-}
