+++ src/main/java/com/blockguard/server/domain/fraud/dto/response/FraudUrlResponse.java
@@ -1,6 +1,6 @@
 package com.blockguard.server.domain.fraud.dto.response;
 
-import com.blockguard.server.domain.fraud.domain.enums.RiskLevel;
+import com.blockguard.server.domain.analysis.domain.enums.RiskLevel;
 import lombok.Builder;
 import lombok.Getter;
